Update client.go to allow the use of paths with the heimdall url
As reported to the polygon team 2 months ago in the Ankr polygon slack, the current code prevents the use of a path with the heimdall url, this was fixed in erigon https://github.com/ledgerwatch/erigon/commit/a3a61701e2cee5b1302ddf2c02dbfcb291522c64 in response to my request by Mark Holt, and it seems this code was just copied from bor where it is incorrect too. This patch fixes that.